spring/Web/SpringMvc0001/src/test/java/com/curso/spring/springmvc0001/web/controladores/PrimerControladorTest.java
/*
 * To change this template, choose Tools | Templates
 * and open the template in the editor.
 */
package com.curso.spring.springmvc0001.web.controladores;

import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.mock.web.MockHttpServletRequest;
import org.springframework.mock.web.MockHttpServletResponse;
import org.springframework.test.context.ContextConfiguration;
import static org.junit.Assert.*;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
import org.springframework.web.servlet.ModelAndView;

/**
 *
 * @author jose maria
 */
@RunWith(SpringJUnit4ClassRunner.class)
@ContextConfiguration("/test-context.xml")
public class PrimerControladorTest {

    public PrimerControladorTest() {
    }

    @BeforeClass
    public static void setUpClass() throws Exception {
    }

    @AfterClass
    public static void tearDownClass() throws Exception {
    }

    @Before
    public void setUp() {
    }

    @After
    public void tearDown() {
    }

    /**
     * Test of handleRequest method, of class PrimerControlador.
     */
    @Test
    public void testHandleRequest() throws Exception {
        System.out.println("testHandleRequest en clase PrimerControladorTest");
        HttpServletRequest request = new MockHttpServletRequest();
        HttpServletResponse response = new MockHttpServletResponse();
        PrimerControlador controlador = new PrimerControlador();
        ModelAndView resultado = controlador.handleRequest(request, response);
        assertEquals("primera", resultado.getViewName());
        assertTrue(resultado.getModel().isEmpty());
    }
}
